+++ reposter-react/src/context/UseAppContext.tsx
@@ -1,0 +1,7 @@
+// UseAppContext.tsx
+import {useContext} from "react";
+import StateContext from "./StateContext";
+
+const useAppContext = () => useContext(StateContext);
+
+export default useAppContext;
